requirements.txt
metaphor-python==0.1.16
pygame==2.1.2
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.27.1
